remove python utils. and move the logger file to the root folder
